requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# HTTP & Web
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Trading
ib-insync==0.9.86
alpaca-trade-api==3.0.2

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration
pyyaml==6.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3



yfinance
